Migrate to new Subscriber types (#434)

+++ src/Frame.h
@@ -180,7 +180,7 @@ class Frame_REQUEST_N {
    *
    * n.b. this is less than size_t because of the Frame encoding restrictions.
    */
-  static constexpr size_t kMaxRequestN = std::numeric_limits<int32_t>::max();
+  static constexpr int64_t kMaxRequestN = std::numeric_limits<int32_t>::max();
 
   Frame_REQUEST_N() = default;
   Frame_REQUEST_N(StreamId streamId, uint32_t requestN)
